prevent round off errors to lower number of theta bins
git-svn-id: https://svn.oss.deltares.nl/repos/xbeach/trunk@1889 86136381-33bf-4961-916e-47ea43f679af

+++ initialize.F90
@@ -151,7 +151,7 @@ contains
        endif
        
        s%dtheta=par%dtheta*degrad
-       s%ntheta=(s%thetamax-s%thetamin)/s%dtheta
+       s%ntheta=nint((s%thetamax-s%thetamin)/s%dtheta)
 
        allocate(s%theta(1:s%ntheta))
        allocate(s%thet(1:s%nx+1,1:s%ny+1,1:s%ntheta))
